requirements.txt
watchdog>=4.0.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.99.0
playwright>=1.38.0
linkedin-api-client>=0.1.0
mcp>=0.3.0
schedule>=1.2.0
requests>=2.31.0
python-dotenv>=1.0.0